Adding local PageHeader component
