+++ src/views/product-downloads-view/components/index.ts
@@ -9,11 +9,13 @@
 import DownloadsSection from './downloads-section'
 import FeaturedTutorialsSection from './featured-tutorials-section'
 import OfficialReleasesSection from './official-releases-section'
+import PageHeader from './page-header'
 import SidecarMarketingCard from './sidecar-marketing-card'
 
 export {
   DownloadsSection,
   FeaturedTutorialsSection,
   OfficialReleasesSection,
+  PageHeader,
   SidecarMarketingCard,
 }
