+++ src/views/product-downloads-view/components/page-header/page-header.module.css
@@ -1,4 +1,4 @@
-.pageHeader {
+.root {
   align-items: center;
   display: flex;
   margin-bottom: 48px;
